+++ qualitycheck-20190115/alibabacloud_qualitycheck20190115/models/_upload_data_sync_response_body.py
@@ -109,10 +109,14 @@ class UploadDataSyncResponseBodyDataResultInfo(DaraModel):
         hand_score_id_list: main_models.UploadDataSyncResponseBodyDataResultInfoHandScoreIdList = None,
         rules: main_models.UploadDataSyncResponseBodyDataResultInfoRules = None,
         score: int = None,
+        tyxm_plus_count: str = None,
+        tyxm_turbo_count: str = None,
     ):
         self.hand_score_id_list = hand_score_id_list
         self.rules = rules
         self.score = score
+        self.tyxm_plus_count = tyxm_plus_count
+        self.tyxm_turbo_count = tyxm_turbo_count
 
     def validate(self):
         if self.hand_score_id_list:
@@ -134,6 +138,12 @@ class UploadDataSyncResponseBodyDataResultInfo(DaraModel):
         if self.score is not None:
             result['Score'] = self.score
 
+        if self.tyxm_plus_count is not None:
+            result['TyxmPlusCount'] = self.tyxm_plus_count
+
+        if self.tyxm_turbo_count is not None:
+            result['TyxmTurboCount'] = self.tyxm_turbo_count
+
         return result
 
     def from_map(self, m: dict = None):
@@ -148,6 +158,12 @@ class UploadDataSyncResponseBodyDataResultInfo(DaraModel):
 
         if m.get('Score') is not None:
             self.score = m.get('Score')
+
+        if m.get('TyxmPlusCount') is not None:
+            self.tyxm_plus_count = m.get('TyxmPlusCount')
+
+        if m.get('TyxmTurboCount') is not None:
+            self.tyxm_turbo_count = m.get('TyxmTurboCount')
 
         return self